refactor(remix-dev/cli/migrate): move `Flags` type into types file
