--- packages/remix-dev/cli/migrate/flags.ts
@@ -1,7 +1,0 @@
-export interface Flags {
-  debug?: boolean;
-  dry?: boolean;
-  force?: boolean;
-  print?: boolean;
-  runInBand?: boolean;
-}
